Added reference to domain classes from Lykke.Domain.Prices.
