+++ src/Core/project.json
@@ -2,7 +2,8 @@
   "version": "1.0.0",
 
   "dependencies": {
-    "Lykke.Common": "1.1.2"
+    "Lykke.Common": "1.1.2",
+    "Lykke.Domain.Prices": "0.3.0"
   },
 
   "frameworks": {
