+++ src/Services/OrderBookService.cs
@@ -2,10 +2,11 @@
 using System.Threading.Tasks;
 using Common;
 using Core.Domain.Assets;
-using Core.Domain.OrderBook;
 using Core.Domain.Settings;
 using Core.Services;
 using Microsoft.Extensions.Caching.Distributed;
+using Lykke.Domain.Prices.Contracts;
+using Lykke.Domain.Prices.Model;
 
 namespace Services
 {
